requirements.txt

openai>=1.40.0
anthropic>=0.33.0
pydantic>=2.7.0
typer>=0.12.3
rich>=13.7.1
tqdm>=4.66.4
pandas>=2.2.2
jinja2>=3.1.3
graphviz>=0.20.3
python-slugify>=8.0.4
streamlit>=1.37.1
requests>=2.32.3
pyyaml>=6.0.2
